+++ db/init.sql
@@ -1,0 +1,4 @@
+-- The development database is created by POSTGRES_DB; creating it here would abort this script.
+
+-- Create test database
+CREATE DATABASE strongmind_ingestion_test;